requirements.txt
astroid==2.0.4
beautifulsoup4==4.6.3
certifi==2018.8.24
cfscrape==1.9.5
chardet==3.0.4
colorama==0.4.0
future==0.17.1
google-images-download==2.4.2
gunicorn==19.9.0
hug==2.4.1
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
line-bot-sdk==1.8.0
mccabe==0.6.1
pylint==2.1.1
requests==2.19.1
selenium==3.14.1
six==1.11.0
urllib3==1.23
wrapt==1.10.11